feat(fw): Add sha256 hash to fixtures
